requirements.txt
Flask
requests
python-dotenv
beautifulsoup4
Pillow
selenium